src/tool_braiinspool/IMPORT/requirements.txt
python-dotenv
pytest

adafruit_io
twilio

git+https://github.com/suchdatums/cdot.git
